ref: Move Dataset enum to own file
Avoids future circular dependency issues as we will be likely needing to
import this class in our Event model in future.
