+++ src/sentry/utils/snuba.py
@@ -5,7 +5,6 @@ from copy import deepcopy
 from contextlib import contextmanager
 from datetime import datetime, timedelta
 from dateutil.parser import parse as parse_datetime
-from enum import Enum, unique
 import os
 import pytz
 import re
@@ -31,6 +30,7 @@ from sentry.net.http import connection_from_url
 from sentry.utils import metrics, json
 from sentry.utils.dates import to_timestamp
 from sentry.snuba.events import Columns
+from sentry.snuba.dataset import Dataset
 
 # TODO remove this when Snuba accepts more than 500 issues
 MAX_ISSUES = 500
@@ -67,15 +67,6 @@ DISCOVER_COLUMN_MAP = {
     for col in Columns
     if col.value.discover_name is not None
 }
-
-
-@unique
-class Dataset(Enum):
-    Events = "events"
-    Transactions = "transactions"
-    Discover = "discover"
-    Outcomes = "outcomes"
-    OutcomesRaw = "outcomes_raw"
 
 
 DATASETS = {
